Fix rare overflow issue in Noise2 and Noise3
In some rare cases, these functions produced wrong outputs which I think
are caused by an overflow.
I couldn't find a bug in Noise4 in my testing, but I applied the same
change there for consistency. Maybe it would have resulted in a bug if I
had let it run the test for a longer period.