Noise: avoid division instruction
While this instruction is optimized to a multiply+shift on instruction
sets with a 32x32=64 bit multiply (smull), this is not the case for
Cortex-M. So it was doing the division in software.

This patch uses a multiply+shift directly. This makes Noise1 over 3.6
times faster on Cortex-M0. The speedup in other noise functions isn't
quite as dramatic, but it's still very substantial.

This change also often improves accuracy slightly.